requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
